Add flag for running in sandbox mode
Fidor's Banking API has a sandbox environment with predefined hostnames. Adding
a flag to the arguments used by this strategy covers most of the use cases where
a consumer would want to change the recipient URLs.

As a result of these changes and moving configuration of the URLs used by the
client inside the `client` method, the `OmniAuth::Fidor::Config` class is
redundant and was removed.

Also corrected the parameters sent during the authorization request and
authorization grant phases.